+++ src/main/php/de/thekid/dialog/OpenMeteo.php
@@ -38,7 +38,14 @@ class OpenMeteo {
       'hourly'     => ['weather_code', 'apparent_temperature'],
     ];
 
-    return $this->endpoint('archive-api')->resource('archive')->post($params, self::URLENCODED)->match([
+    // There's a 5-day delay in the historical data; forecasts are provided for 7 days
+    if ($start->getTime() >= time() - 604800) {
+      $resource= $this->endpoint('api')->resource('forecast');
+    } else {
+      $resource= $this->endpoint('archive-api')->resource('archive');
+    }
+
+    return $resource->post($params, self::URLENCODED)->match([
       200 => fn($r) => $r->value(),
       400 => fn($r) => throw new IllegalArgumentException($r->content()),
     ]);